requirements.txt
dnspython>=2.2.0
PyMySQL>=1.0.2
PyYAML>=5.4.1
